+++ client/components/Header/MenuButton.tsx
@@ -1,0 +1,57 @@
+import React, { MouseEventHandler } from 'react';
+import { testWrap } from '../testWrap/testComp';
+import './menuButton.css';
+
+interface IBar {
+  wide?: boolean;
+  classNames: string;
+}
+
+function Bar({ wide = false, classNames }: IBar): JSX.Element {
+  return <div className={`${wide ? 'w-[35px]' : 'w-[26px]'} bar ${classNames}`} />;
+}
+
+interface IHamburgerC {
+  showClose: boolean;
+}
+
+function HamburgerC({ showClose }: IHamburgerC): JSX.Element {
+  return (
+    <div
+      aria-hidden
+      style={{
+        width: '35px',
+        height: '16px',
+        position: 'relative',
+      }}
+    >
+      <Bar classNames={cls('top', showClose)} />
+      <Bar wide classNames={cls('middle', showClose)} />
+      <Bar classNames={cls('bottom', showClose)} />
+    </div>
+  );
+}
+
+const Hamburger = testWrap(HamburgerC, 'hamburger');
+
+interface IMenuButton {
+  onClick: MouseEventHandler<HTMLButtonElement>;
+  showClose: boolean;
+}
+
+export function MenuButton({ onClick, showClose }: IMenuButton): JSX.Element {
+  return (
+    <button
+      type="button"
+      onClick={onClick}
+      className="menu-button outline-1 outline-thmPrimary p-[10px] outline-offset-[-5px]"
+    >
+      <span className="sr-only">{showClose ? 'timer' : 'settings'}</span>
+      <Hamburger showClose={showClose} />
+    </button>
+  );
+}
+
+function cls(classes: string, showClose: boolean): string {
+  return showClose ? `${classes} showClose` : classes;
+}
