+++ client/components/index.ts
@@ -4,7 +4,7 @@ export * from './Checkbox';
 export { default as ErrorBoundary } from './ErrorBoundary/ErrorBoundary';
 export * from './Inspector';
 export * from './LineBreak';
-export * from './MenuButton';
+export * from './Header/MenuButton';
 export * from './testWrap/testComp';
 export * from './PageManager/PageManager';
 export * from './Pomodoro/Pomodoro';
